server/requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrdict==2.0.1
attrs==22.2.0
Babel==2.10.2
bce-python-sdk==0.8.73
boto3==1.24.0
botocore==1.27.0
cachetools==5.2.0
certifi==2022.5.18.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
cssselect==1.1.0
cssutils==2.4.2
cycler==0.11.0
Cython==0.29.30
datefinder==0.7.3
decorator==5.1.1
distlib==0.3.4
dnspython==2.2.1
et-xmlfile==1.1.0
easyocr==1.6.2
fastapi==0.95.0
filelock==3.7.1
filetype==1.2.0
flake8==4.0.1
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
flatbuffers==23.1.4
fonttools==4.33.3
frozenlist==1.3.3
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
huggingface-hub==0.8.1
humanfriendly==10.0
identify==2.5.1
idna==3.3
ImageHash==4.3.1
imageio==2.19.3
imgaug==0.4.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==15.0.6.1
llvmlite==0.39.1
lmdb==1.3.0
lxml==4.9.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.6.1
mpmath==1.3.0
multidict==6.0.4
networkx==2.6.3
nltk==3.7
nodeenv==1.6.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.14.1
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.2
paddleocr==2.5.0.3
paddlepaddle==2.3.0
pandas==1.3.5
pdf2image==1.16.0
pika==1.2.1
Pillow==9.5.0
pillow-heif==0.10.1
platformdirs==2.5.2
pooch==1.7.0
poppler-utils==0.1.0
pre-commit==2.19.0
premailer==3.10.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pycodestyle==2.8.0
pycryptodome==3.14.1
pydantic==1.10.7
pyflakes==2.4.0
PyMatting==1.1.8
pymongo==4.1.1
pyparsing==3.0.9
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.12.2
python-multipart==0.0.6
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.7.25
rembg==2.0.32
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.10.1
Shapely==1.8.2
shellcheck-py==0.8.0.4
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.12.1
toml==0.10.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.0
transformers==4.21.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.21.1
virtualenv==20.14.1
visualdl==2.2.3
watchdog==3.0.0
Werkzeug==2.1.2
wordtodigits==1.0.2
wrapt==1.14.1
xlrd==2.0.1
yarl==1.8.2
zipp==3.8.0
